register and login with token, configuration, fix userRepo, pom xml add jwt dependency

+++ src/main/java/com/example/sapar/repositories/UserRepository.java
@@ -2,13 +2,12 @@ package com.example.sapar.repositories;
 
 import com.example.sapar.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 
     boolean existsByUsername(String username);
